requirements.txt
Flask==2.0.3
gunicorn==20.1.0
python-dotenv==0.19.2
firebase-admin